requirements.txt
streamlitpython-dotenvsupabasepandasplotly  